thirdparty: download the torch library
On the one hand, the TO project need a specific version:
we can't use the default debian version package management.
On the other hand, the docker continuous integration can't address
the standalone dependencies.
So we manage directely this specific dependency
downloading it.

Manage the CMakeLists.txt include path.

Copy the "include" source directory
into the build directory
to be able to handle it into the running docker image
and run the "exchange" binary without keep a piece
of "source".

Use CMake 3.11 instead of the deprecated 3.0 version (compatibilty issues
are declared between cx17 and cmake < 3.11).